Split FretboardModel and FretboardModelExt
